+++ src/E5R.Architecture.Infrastructure/Abstractions/ICrossCuttingRegistrar.cs
@@ -2,12 +2,13 @@
 // This file is a part of E5R.Architecture.
 // Licensed under the Apache version 2.0: https://github.com/e5r/manifest/blob/master/license/APACHE-2.0.txt
 
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace E5R.Architecture.Infrastructure.Abstractions
 {
-    public interface IDIRegistrar
+    public interface ICrossCuttingRegistrar
     {
-        void Register(IServiceCollection services);
+        void Register(IServiceCollection services, IConfiguration configuration);
     }
 }